+++ test/blas/fp16/level1/check_hnrm2_fp32.c
@@ -54,9 +54,9 @@ void check_hnrm2( bool *ok) {
     for ( int i = 0; i < NTEST; i++) {
         lpf_float16_t result = from_u16(LPF_GLOBAL(hnrm2_fp32, HNRM2_FP32)(&n[i], data[i], &incx[i]));
         if ( diff_fp16(result, hnrm2_expected[i]) < __FLT16_EPSILON__ * 10) {
-            printf("HNRM2_FP32   -- OK   -- Testcase %2d: N = %3d, INCX = %3d, Result = %10f, Expected = %10f\n", i+1, n[i], incx[i], (float)result, (float)hnrm2_expected[i]);
+            printf("HNRM2_FP32   -- OK   -- Testcase %2d: N = %3d, INCX = %3d, Result = %10f, Expected = %10f\n", (int)i+1, (int)n[i], (int)incx[i], (float)result, (float)hnrm2_expected[i]);
         } else{
-            printf("HNRM2_FP32   -- FAIL -- Testcase %2d: N = %3d, INCX = %3d, Result = %10f, Expected = %10f\n", i+1, n[i], incx[i], (float)result, (float)hnrm2_expected[i]);
+            printf("HNRM2_FP32   -- FAIL -- Testcase %2d: N = %3d, INCX = %3d, Result = %10f, Expected = %10f\n", (int)i+1, (int)n[i], (int)incx[i], (float)result, (float)hnrm2_expected[i]);
             *ok = false;
         }
     }
