+++ test/blas/fp16/level1/check_hshdot.c
@@ -67,9 +67,9 @@ void check_hshdot( bool *ok) {
     for ( int i = 0; i < NTEST; i++) {
         lpf_float16_t result = from_u16(LPF_GLOBAL(hshdot, HSHDOT)(&n[i], &b, x[i], &incx[i], y[i], &incy[i]));
         if ( diff_fp16(result, hshdot_expected[i]) < __FLT16_EPSILON__ * 10) {
-            printf("HSHDOT -- PASS -- Testcase %2d: N = %3d, INCX = %3d, INCY = %3d, Result = %10f, Expected = %10f\n", i+1, n[i], incx[i], incy[i], (float)result, (float)hshdot_expected[i]);
+            printf("HSHDOT -- PASS -- Testcase %2d: N = %3d, INCX = %3d, INCY = %3d, Result = %10f, Expected = %10f\n", (int)i+1,(int) n[i],(int) incx[i],(int) incy[i], (float)result, (float)hshdot_expected[i]);
         } else{
-            printf("HSHDOT -- FAIL -- Testcase %2d: N = %3d, INCX = %3d, INCY = %3d, Result = %10f, Expected = %10f\n", i+1, n[i], incx[i], incy[i], (float)result, (float)hshdot_expected[i]);
+            printf("HSHDOT -- FAIL -- Testcase %2d: N = %3d, INCX = %3d, INCY = %3d, Result = %10f, Expected = %10f\n", (int)i+1,(int) n[i],(int) incx[i],(int) incy[i], (float)result, (float)hshdot_expected[i]);
             *ok = false;
         }
     }
